Clean up app and fix build
