+++ .circleci/config.yml
@@ -1,22 +1,19 @@
 version: 2.1 
-orbs:
-  browser-tools: circleci/browser-tools@1.1
 jobs:  # basic units of work in a run
   build:  # runs not using Workflows must have a `build` job as entry point
     parallelism: 1  # run only one instance of this job
     docker:  # run the steps with Docker
-      - image: cimg/elixir:1.12.2-browsers  # ...with this image as the primary container; this is where all `steps` will run
+      - image: cimg/elixir:1.12.2 # ...with this image as the primary container; this is where all `steps` will run
         auth:
           username: mydockerhub-user
           password: $DOCKERHUB_PASSWORD  # context / project UI env-var reference
         environment:  # environment variables for primary container
           MIX_ENV: test
-          TERM: xterm
+          TERM: xterm-256color
 
     working_directory: ~/app  # directory where steps will run
 
     steps:  # commands that comprise the `build` job
-      - browser-tools/install-browser-tools
       - checkout  # check out source code to working directory
       - run: sudo apt update
       - run: sudo apt install python3  # install python for ex_termbox build
