--- config/runtime.exs
@@ -1,5 +1,0 @@
-import Config
-
-config :logger,
-  level: :error,
-  truncate: 4096
