+++ mix.exs
@@ -14,7 +14,8 @@ defmodule Chaffinch.MixProject do
   # Run "mix help compile.app" to learn about applications.
   def application do
     [
-      extra_applications: [:logger]
+      extra_applications: [:logger],
+      mod: {Chaffinch, []}
     ]
   end
